Add -qpause and -qresume, add more ayah reciters, rewrite -reciters and fix some bugs.
This commit adds  and , which pause and resume the bot respectively. It also switches the ayah-by-ayah recitation provider to mp3quran.net, unlocking more reciters for . In addition,  has been rewritten toprovide a paginated list of reciters for each  subcommand. Finally, error handling has been improved, which fixes some bugs in the process.